Fix code analysis issues and add aurora effect to VIP banner
- Remove unused imports in bookshelf_page_tabs and energy_records_page
- Remove unnecessary type cast in app_swipe_tab_switcher
- Remove unused child parameter in aurora_background
- Use initializing formal in help_feedback_cubit
- Replace static gradient with aurora animation in VIP banner using pink-purple color scheme

+++ lib/shared/components/app_swipe_tab_switcher.dart
@@ -79,7 +79,7 @@ class _AppSwipeTabSwitcherState extends State<AppSwipeTabSwitcher> {
     if (children == null) return _buildGestureOnly(widget.child!);
     if (children.isEmpty) return const SizedBox.shrink();
     if (!widget.enabled || children.length <= 1) {
-      final index = widget.selectedIndex.clamp(0, children.length - 1) as int;
+      final index = widget.selectedIndex.clamp(0, children.length - 1);
       return children[index];
     }
